type errors fixed

+++ src/main/g8/src/main/scala/$package$/models/GreetingEntity.scala
@@ -1,5 +1,5 @@
 package $package$.models
 import slick.driver.H2Driver.api._ 
 
-case class GreetingEntity(greeting: String, id: Int = 0)
+case class GreetingEntity(greeting: String, id: Long = 0)
